Add the value of the contact to the view state.
Write code to add the contact (text) to the view state. This particular
implementation adds the value of the contact field to the state (with
the key `:text`). In addition, it renders the component by extracting
the `:text` component from the state and setting the `:value` component
of the (contact) input field to that extracted value.

A problem results. Whenever I type a character, the component detects
that new character and adds it to the input field. The component that
detects that change and re-renders itself resulting in setting the value
of the input field to an empty string. As a result, one appears not to
be able to enter any characters.

To be fixed with the next change.