Repair the "cannot edit contact" error.
Change the code to repair the error that results in *not* being able to
edit the contact field. This change involves adding a function to handle
all changes to the contact field. This handler sets the state of the
contact field (tagged by `:text` in the app-state) to the *new* value of
the (contact) input field. Setting the app-state results in Om/React
re-rendering the input field with the changed value of the contact.
(Note that this mechanism is the same mechanism described by Martin
Fowler in, I believe, "Patterns of Enterprise Architecture," to describe
how to correctly synchronize models and views.)