Change implementation of clearing contact.
Change the implementation of clearing the newly added contact from
directly manipulating the value property of the (contact) input field to
manipulating the :text component of the application state.

This change moves ClojureScript code from "Windows UI code that directly
manipulates the UI controls" to "MVC-style" code that changes the model
(`app-state`) which results in changing the view.

Amazing.