Code from "Higher Order Components" (part 1).
Write the code from the first part of the section "Higher Order
Components." This code supports displaying a page with the heading
"Registry"; however, it does not support display different views for
students and professors.

Basically, I captured this commit because the next (sub-)section asks me
to change the rendering of the registry component. I wanted to capture
the code before making this change.