Code from "Higher Order Components" (part 2).
Write code from the second part of "Higher Order Components." This
commit results displaying a list of students and professors and actually
displaying **different** results for students than for professors.
Specifically, for students, this commit displays the name of the
student, but for professors, it displays their name but also the classes
that they teach.